+++ src/main/java/com/github/wubuku/sui/bean/TransferObjectEvent.java
@@ -1,5 +1,7 @@
 package com.github.wubuku.sui.bean;
 
+import java.math.BigInteger;
+
 /**
  * From TypeScript definition:
  * <p>
@@ -22,7 +24,7 @@ public class TransferObjectEvent {
     private ObjectOwner recipient;
     private String objectType;
     private String objectId;
-    private Long version;
+    private BigInteger version;
 
     public String getPackageId() {
         return packageId;
@@ -72,11 +74,11 @@ public class TransferObjectEvent {
         this.objectId = objectId;
     }
 
-    public Long getVersion() {
+    public BigInteger getVersion() {
         return version;
     }
 
-    public void setVersion(Long version) {
+    public void setVersion(BigInteger version) {
         this.version = version;
     }
